requirements (5).txt
# https://pypi.org/project/flask
Flask==3.0.2

# https://pypi.org/project/gunicorn
gunicorn==21.2.0

# https://pypi.org/project/google-cloud-storage
google-cloud-storage==2.19.0

# https://pypi.org/project/flask-session
flask-session==0.4.0

# Additional dependencies
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
Pillow==10.0.1
proto-plus==1.25.0
protobuf>=3.20.2,<6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.3.0
Werkzeug==3.1.3
Flask==3.0.2
firebase-admin==6.6.0
google-cloud-storage==2.19.0
gunicorn==21.2.0